# optimize int binning