* combined the AbstractFloat and Integer implementations. 
* swapped the `+0.5`/`round` for just `floor` based on what I saw in https://github.com/joshday/OnlineStats.jl/blob/master/src/stats/histograms.jl. Though if I benchmark their hist, it's ~3-4x slower than FHist, probably related to them not using `@inline`. 3-4x is the speedup I observed when adding `@inline` in front of `unsafe_push!()`
* the `ifelse` chain evidently wasn't even triggering simd, so the implementation is simpler now.


`-3:0.1:3` is slightly faster, and `-3:3` is even faster (and 8x faster than master).


Before (`master`):
```julia
julia> a = randn(10^6);

julia> @benchmark Hist1D(a, -3:0.1:3)
BechmarkTools.Trial: 1142 samples with 1 evaluations.
 Range (min … max):  3.839 ms …   6.746 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     4.313 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   4.372 ms ± 239.986 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%
 Memory estimate: 1.98 KiB, allocs estimate: 9.

julia> @benchmark Hist1D(a, -3:3)
BechmarkTools.Trial: 230 samples with 1 evaluations.
 Range (min … max):  20.159 ms …  25.909 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     21.708 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   21.794 ms ± 779.543 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%
 Memory estimate: 592 bytes, allocs estimate: 8.
```

After:
```julia
julia> a = randn(10^6);

julia> @benchmark Hist1D(a, -3:0.1:3)
BechmarkTools.Trial: 1244 samples with 1 evaluations.
 Range (min … max):  3.550 ms …   9.857 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     3.895 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   4.006 ms ± 412.422 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%
 Memory estimate: 1.98 KiB, allocs estimate: 9.

julia> @benchmark Hist1D(a, -3:3)
BechmarkTools.Trial: 1838 samples with 1 evaluations.
 Range (min … max):  2.465 ms …   4.494 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     2.633 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.711 ms ± 204.492 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%
 Memory estimate: 592 bytes, allocs estimate: 8.
```